requirements.txt
celery
eventlet
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
lxml
parsel
PyExecJS
PyMySQL
requests
tenacity
WTForms
gunicorn
opencv-python
numpy
beautifulsoup4
pillow
selenium